+++ src/xclim/core/collection.py
@@ -37,10 +37,7 @@ details on each.
       <identifier>:  # The actual indicator identifier will be prepended by the module name.
         # From which Indicator to inherit
         base: <base indicator class>  # Defaults to module-wide base class
-                                      # If the name startswith a '.', the base class is taken from the current module
-                                      # (in the `bases` section or from another indicator declared _above_)
-                                      # Available indicators are listed in `xclim.core.indicator.registry` and
-                                      # other base classes in `xclim.core.indicator.base_registry`.
+                                      # See :ref:`Base class specification` below.
 
         # General metadata, usually parsed from the `compute`s docstring when possible.
         realm: <realm>  # defaults to module-wide realm. One of "atmos", "land", "seaIce", "ocean".
@@ -58,12 +55,12 @@ details on each.
         context: <context> # A unit context enabled during the conversion of the compute's output to the requested units
 
         # Compute function
-        compute: <function name>  # Referring to a function in `compute` module
-                                  # (xclim.compute.generic or xclim.compute)
-                                  # Or to a function declared in the mapping passed to the collection constructor.
+        compute: <function name>  # See :ref:`Compute function specification`, below.
+
         input:  # When "compute" is a generic function, this is a mapping from argument name to the expected variable.
                 # It will change the expected name of the variable as well as its units/dimensionality.
                 # Can refer to a variable declared in the `variables` section above or in `xclim.core.VARIABLES`.
+                # See also :ref:`Inputs` below.
           <var name in compute> : <variable official name>
           ...
 
@@ -84,6 +81,37 @@ All fields are optional. Other fields found in the yaml file will trigger errors
 When a module is built from a yaml file, the yaml is first validated against the schema (see xclim/data/schema.yml)
 using the YAMALE library (:cite:p:`lopker_yamale_2022`). See the "Extending xclim" notebook for more info.
 
+Base class specification
+^^^^^^^^^^^^^^^^^^^^^^^^
+There are multiple ways to specify a base class when defining an indicator. In priority order:
+
+- If ``base`` starts with a '.' (ex: `.RXXp`), the base class is taken from the current module.
+    + It is first searched in `bases` section.
+    + If not found, it is searched as another indicator declared _above_ the current definition.
+- The name is searched in the base class registry, :py:data:`xclim.core.indicator.base_registry` (example: ``Daily``).
+- The name is searched in the indicator registry, :py:data:`xclim.core.indicator.registry` (example: ``prcptot``).
+- If ``base`` contains a '.' :
+    + If the first element is one of xclim's indicators submodules (ex: ``atmos.precip_accumulation``), that indicator
+      is used as a base. Any submodule of ``xclim.indicators`` are possible.
+    + Otherwise, that path is loaded with python's normal import mechanism.
+      (ex: ``mymodule.submod.MyIndicator`` isloaded as ``from mymodule.submod import MyIndicator``).
+
+If the field ``base`` is not given, it defaults to the module-wide ``base``, which itself defaults to
+:py:class:`xclim.core.indicator.Daily``.
+
+Compute function specification
+^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
+Similar to the ``base`` field, there are multiple ways to refer to a compute function in the ``compute`` field.
+In priority order:
+
+- If a module or mapping of compute functions was passed to :py:meth:`IndicatorCollection.from_yaml`,
+  the name is searched there (ex: `extreme_precip_accumulation_and_days`).
+- The name is searched in :py:mod:`xclim.compute.generic` (ex: ``statistics``).
+- The name is searched in `:py:mod:`xclim.compute` (ex: ``corn_heat_units``).
+  It may contain a '.' to denote a submodule (ex: ``generic.statistics``).
+- Otherwise, it is loaded with python's normal import mechanism
+  (ex: ``mymodule.submod.my_function`` is loaded as ``from mymodule.submod import my_function``).
+
 Inputs
 ^^^^^^
 As xclim has strict definitions of possible input variables (see :py:data:`xclim.core.VARIABLES`),
@@ -95,6 +123,7 @@ from __future__ import annotations
 
 import warnings
 from collections.abc import Callable
+from importlib import import_module
 from os import PathLike
 from pathlib import Path
 from types import ModuleType
@@ -341,18 +370,38 @@ class IndicatorCollection(dict):  # numpydoc ignore=PR01
 
     @staticmethod
     def _find_base_class(name, mapping, bases):
-        if name.startswith("."):
-            # A point means the base has been declared above.
-            if name[1:] in bases:
-                base = bases[name[1:]]
+        try:
+            if name.startswith("."):
+                # A starting dot means the base has been declared above.
+                if name[1:] in bases:
+                    base = bases[name[1:]]
+                else:
+                    base = mapping[name[1:]].__class__
+            elif name in base_registry:
+                base = base_registry[name]
+            elif name in registry:
+                base = registry[name].__class__
+            elif "." in name:
+                # A dot not at the start means a qualified name either relative to xclim.indicators or full
+                if name.count(".") == 1 and name.split(".")[0] in [
+                    "atmos",
+                    "land",
+                    "seaIce",
+                    "convert",
+                    "generic",
+                    "cf",
+                    "anuclim",
+                    "icclim",
+                ]:
+                    name = f"xclim.indicators.{name}"
+                modname, indname = name.rsplit(".", 1)
+                base = getattr(import_module(modname), indname)
+                if isinstance(base, Indicator):
+                    base = base.__class__
             else:
-                base = mapping[name[1:]].__class__
-        elif name in base_registry:
-            base = base_registry[name]
-        elif name in registry:
-            base = registry[name].__class__
-        else:
-            raise ValueError(f"Can't find requested base class {name}.")
+                raise KeyError(name)
+        except Exception as err:
+            raise ValueError(f"Can't find base class {name}.") from err
         return base
 
     @staticmethod
@@ -360,17 +409,23 @@ class IndicatorCollection(dict):  # numpydoc ignore=PR01
         func = None
         if computes is not None:
             func = getattr(computes, name, None)
-        if func is None:
-            if hasattr(computes, "__getitem__") and name in computes:
+            if func is None and hasattr(computes, "__getitem__") and name in computes:
                 func = computes[name]
-            elif "." in name:
+        if func is None:
+            if "." in name:
                 modname, name = name.split(".")
                 submod = getattr(xclim.compute, modname, None)
                 func = getattr(submod, name, None)
             else:
                 func = getattr(xclim.compute.generic, name, getattr(xclim.compute, name, None))
+        if func is None and "." in name:
+            modname, funcname = name.rsplit(".", 1)
+            try:
+                func = getattr(import_module(modname), funcname)
+            except ModuleNotFoundError as err:
+                raise_warn_or_log(err, "log", msg=f"Failed importing {funcname} from {modname}.")
         if func is None:
-            raise ValueError(f"Can't find requested compute function {name}.")
+            raise ValueError(f"Can't find compute function {name}.")
         return func
 
     def __dir__(self):